pyproject.toml
[tool.poetry]
name = "cowrywise_uuid_api"
version = "0.1.0"
description = ""
authors = ["Ajawu David <ajawudavid@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest-mypy = "^0.8.1"
djangorestframework-stubs = "^1.4.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

